Add JWKS endpoint (#26)
This PR adds a public endpoint for getting the session signing keys for
a project in JWKS format. This will be used by clients to authenticate
sessions from their backend.

I opt to implement this in a new "oauthservice" HTTP server. I do this
because 1) this endpoint is weird; it doesn't want any authentication at
all, and probably wants to be specially cached by a CDN down the line,
and 2) it fits most naturally as part of an overall project-level OAuth
authorization server.

I opt to change the store layer method for getting public keys to return
a set of them, and make the store layer be in charge of converting keys
to their standardized representation.