Add CI checks for `make proto` and `make queries` (#52)
This PR adds CI checks that ensure that `make proto` and `make queries`
don't result in any `git diff` changes.

Technically, it's not `make queries` but its underlying `sqlc`
invocation that is checked. There's so pgformat stuff in `make queries`
that I don't want to wrap into what is otherwise a super fast CI check.
The existing pgformat-lint CI check is still there to check for bad
changes to that file.