Handle failed challenges & create verified emails (#53)
This updates the schema for the `email_verification_challenges` table to
persist state on `intermediate_sessions` only, and updates the
`verified_emails` table to include `google_hosted_domain` and
`microsoft_tenant_id`.

This also adds logic to handle verification checks and revoke failures
as well as create a new `verified_emails` record on success.

There's also an additional
`/intermediate/v1/issue-email-verification-challenge` endpoint added to
support retrying of email verification. In the future, we'll need to add
some constraint checks to ensure this is capped to avoid abuse.