[OPEN-45] Add GetProject, UpdateProject to backend api (#73)
This PR adds a CRUD layer for projects to the backend API.

ListProjects is removed entirely. It never made sense to have exist.
CreateProject is removed entirely as well, but may reincarnate when we
do onboarding work.

GetProject and UpdateProject don't ask for a project_id parameter; it's
implied via `projectid.ProjectID(ctx)`, which is where the interesting
logic will live. It will be projectid's job to ensure alignment between
authentication methods and the project they think they're operating on.

This PR adds display names to projects. It's a plain old mutable
attribute on projects.

The dogfood project is named "OpenAuth Dogfood". Project-backing
organizations are named after the ID of the project they back. That
logic is happening in openauthctl/bootstrap.

As part of this work, I'm doing a pass over removing comments in proto,
as well as what I judge to be unnecessary comments in the go code. For
the proto comments, we'll figure out the tone and generic copy we want
to use later, because that'll be in large part determined by what makes
most sense in the online docs and in the SDKs.

I don't think UpdateProject should be callable from project API keys.
That should be a human-only endpoint. I've created a ticket to track
that work, which I'm not implementing in this PR.

Demo of getting, then updating, then getting again a project:

```
$ curl -H "X-TODO-OpenAuth-Project-ID: project_3jzjqbc6nogleg3c8gt9f3mbt" -H "content-type: application/json" -H "authorization: Bearer openauth_secret_key_er5rdn41czbgb78htvcqen2za" localhost:3001/backend/v1/project | jq

{
  "project": {
    "id": "project_3jzjqbc6nogleg3c8gt9f3mbt",
    "displayName": "OpenAuth Dogfood"
  }
}

$ curl -H "X-TODO-OpenAuth-Project-ID: project_3jzjqbc6nogleg3c8gt9f3mbt" -H "content-type: application/json" -H "authorization: Bearer openauth_secret_key_er5rdn41czbgb78htvcqen2za" localhost:3001/backend/v1/project -XPATCH -d '{ "displayName": "OpenAuth Dogfood!!!" }'
{"project":{"id":"project_3jzjqbc6nogleg3c8gt9f3mbt","displayName":"OpenAuth Dogfood!!!"}}

$ curl -H "X-TODO-OpenAuth-Project-ID: project_3jzjqbc6nogleg3c8gt9f3mbt" -H "content-type: application/json" -H "authorization: Bearer openauth_secret_key_er5rdn41czbgb78htvcqen2za" localhost:3001/backend/v1/project | jq

{
  "project": {
    "id": "project_3jzjqbc6nogleg3c8gt9f3mbt",
    "displayName": "OpenAuth Dogfood!!!"
  }
}
```