Fixes to make intermediate-issued sessions work with the backend (#80)
This PR makes intermediate-issued sessions work with the backend
interceptor.

I am not making any claims as to whether the sessions module or the
backend interceptor are good at this time. I suspect we'll end up making
changes to both in the future. Instead, I'm just trying to make these
things talk to each other succesfully.