[OPEN-75] Basic cookie management (#96)
This PR adds support for cookies on both intermediate session and
sessions.

In order to make this work locally, an `nginx` container was added to
the `docker-compose.yaml` which is running all services on sub-domains
of `tesseral.local`. The subdomains are `api.tesseral.local`,
`app.tesseral.local`, and `ui.tesseral.local`, respectively.

**_This requires an update to both the local `.env` file and your
machine's `/etc/hosts` file. An example `etc.hosts.example` file has
been included in this commit as well._**

To verify that cookies were working in both cases, the `Whoami` service
function has been implemented for the frontend service, and the
`/session-info` page has been stubbed out to render email to the screen
in the ui after session exchange.

![Screenshot 2025-01-09 at 12 25
45 PM](https://github.com/user-attachments/assets/ed724109-ba8a-4cca-be9b-e771c6bc44fc)