Add primitive useAccessToken hook to app (#107)
This PR adds a primitive, working useAccessToken hook to the app. I put
this PR up in this state because it contains bugfixes relevant to other
parts of the system -- in particular, refresh tokens were not being
stored properly, because their stringified representation, instead of
the underlying UUID bytes, were being SHA'd.

This PR also makes the GetAccessToken endpoint skip authz checks; it
authenticates its refresh token parameter on its own.

The useAccessToken hook does not cache any data. That work will happen
in a future PR, alongside work to compose on top of what's available
with an access token on hand.


![screenshot-2025-01-12-15-58-07](https://github.com/user-attachments/assets/f1975f25-9c4d-4231-853b-cc67c8c9f956)

---------

Co-authored-by: Blake Williams <blakethewilliams@gmail.com>