[OPEN-86] [OPEN-87] Custom project domains and projectID sniffing (#109)
# Changes

## Custom project domains

This PR adds support for the `custom_domain` column on projects. In
order to accomplish this, it does the following:
- Updates the `projects` table to include a new `custom_domain varchar`
column
- Updates the backend `UpdateProject` query to support `custom_domain`
- Updates the `parseProject` functions in all services to include
`CustomDomain` in the response

## Project ID sniffing

It also adds the logic required to determine projectIDs from the domain.
In order to accomplish this it does the following:
- Adds a new `GetProjectIDByDomain` query to all service-level stores
except the OAuth service
- Updates the authn interceptors on the `frontend`, `intermediate`,
`saml`, and `scim` services to check for project ID by domain
- In the case of auto-generated subdomains, this is done by using a
regular expression to match the
`project-XXXXXXX.{config.AuthAppsRootDomain}` string and extracting the
project ID from the host header
- In the case of custom domains, this is done by utilizing the
`GetProjectIDByDomain` query to check for a project with the value of
the host header in the its `custom_domain` column

## Basic backend protections

This PR also adds a check to disallow backend requests on any hosts
other than the api host itself (now in an environment variable) and the
app auth host (also in an environment variable).

## Additional environment variables

This PR also updates `cmd/api/main.go` to consume a few new environment
variables:
- `API_HOST` - used to enforce backend access rules; on localhost, this
is `api.tesseral.example.com`
- `API_AUTH_APPS_ROOT_DOMAIN` - used to perform regex matches in authn
middleware; on localhost, this is `tesseral.example.app`
- `API_DOGFOOD_AUTH_DOMAIN` - used to enforce backed access rules; on
localhost, this is `auth.app.tesseral.example.com`

These values are then passed down to their respective middleware as
needed.

# Testing

In order to test this I first updated a project in the database to use
the `auth.app.tesseral.example.com` domain as its custom domain.

With that set, I accessed `auth.app.tesseral.example.com` to ensure that
the projectID was resolving through the entirety of the login flow and
additionally visited the app at `app.tesseral.example.com` to ensure
that the projectID was populated correctly on its requests.

I then performed the same round of tests on
`project-XXXXXXX.tesseral.example.app` to perform projectID sniffing
worked on auto-generated subdomains.