[OPEN-101] project UI settings (#136)
# Overview

This PR adds support for getting and updating project UI settings in
`backend` service and getting project UI settings in `intermediate`
service. In order to accomplish this, it adds a new
`project_ui_settings` table to the database, and updates the `backend`
and `intermediate` protos to include the necessary RPCs.

## Project UI Settings

The structure of project UI settings is as follows:
```
{
    id: string
    projectId: string
    createTime: timestamp
    updateTime: timestamp
    detectDarkModeEnabled: boolean
    darkModeLogoUrl: string
    faviconUrl: string
    logoUrl: string
    primaryColor: string
}
```

## Asset management

The `UpdateProjectUISettings` rpc responds with a reflection of current
settings as well as `darkModeLogoPresignedUploadUrl`,
`faviconPresignedUploadUrl`, and `logoPresignedUploadUrl` properties.
These pre-signed upload URLs can be used by the client to upload the
necessary assets for the current set of project UI settings.

## Local S3 Mocking

In order to support a local run of this functionality, we're now using
the [adobe/S3Mock](https://github.com/adobe/S3Mock) docker image in our
`docker-compose.yaml` to run a functional mock s3 server locally. This
writes data to the local file system at `.local/s3`. This requires two
new `API_S3_BUCKET_NAME` and `API_S3_ENDPOINT_RESOLVER_URL` environment
variables (only required locally).

Additionally, the local nginx config has been updated to point
`tesseralusercontent.example.com` to the mock s3 bucket running in
docker. This can serve as a local proxy for a properly configured bucket
in production. This host is used in the `API_USER_CONTENT_URL`
environment variable to build asset URLs.

## Basic UI implementation

The UI now includes a query to the `GetProjectUISettings` rpc on load,
which is currently only used to control dark mode behavior, but will
include full customization in a future PR. This PR was more about the
server-side implementation, so the UI implementation is mostly just to
verify the service layer.

The UI implementation also includes some basic failover logic for logos
to use the default tesseral image if the URL provided in the project UI
settings fails.

# Testing

To test the basics of this behavior, I ran bootstrap (which now creates
the `project_ui_settings` record for the dogfood project) and updated
the `detect_dark_mode_enabled` to `false`. I was then able to visit the
UI and see this without changing any settings on my machine (my machine
is in dark mode).

![Screenshot 2025-01-16 at 9 19
19 AM](https://github.com/user-attachments/assets/53571065-c26a-46a9-84b4-91df44eb5135)

After changing `detect_dark_mode_enabled` to `true` again, I'm able to
visit the UI and see this without changing any settings on my machine.

![Screenshot 2025-01-16 at 9 19
07 AM](https://github.com/user-attachments/assets/cd106898-a4f2-4638-83bb-250b6c36698d)