Add stacked concept for ListOrganizationsPage (#146)
The main subtlety here is that there's no natural details view for this
page. So instead I'm thinking we do an edge-to-edge card/table? Only
cards enjamb the fold gracefully.

When we add a button for creating a new org, I figure it goes at the
level of the page title, but on the right.


![screenshot-2025-01-17-15-44-15](https://github.com/user-attachments/assets/6cd5c605-f6fc-468e-9bd3-a3d2bb48130c)