Centralize access token issuance into accesstoken.Issuer (#153)
This PR replaces all the places where we issue access tokens with a
single centralized module. The claims on these JWTs are now in
common.proto.

There's some code in intermediate store that I just comment out -- Go is
upset that we load an org without doing something with it, but a later
PR will need that org loaded up anyway, in order to enforce org-level
auth requirements.