[OPEN-127] rename get access token (#163)
This PR renames the `GetAccessToken` rpc and all of its referencing
functions to `Refresh`. It also updates the REST route to
`/frontend/v1/refresh` and updates client side references.

Additionally, a migration conflict was blocking testing, so I updated
the name of the latest migration to fix the sequential ordering of our
migrations.