Fix backend dogfood session authn (#170)
This PR fixes the logic the backend interceptor uses to determine what
project you're working on. It's now based on looking up the organization
the user belongs to, and finding the project that organization is a
stand-in for. OrganizationID is removed from the context data, because
it's misleading to look at.

To make this PR work in a way that isn't a distinction without a
difference, I add two new orgs to seed.sql.

Here's what user1 at company2 sees (after I manually issue a
CreateOrganization call using their access token):


![screenshot-2025-01-24-15-47-15](https://github.com/user-attachments/assets/04b3cf70-0c66-4487-b261-a53d555a0190)

Here's what root at the dogfood project sees:


![screenshot-2025-01-24-15-48-59](https://github.com/user-attachments/assets/54df9a83-c63b-4e75-ad98-e859e4fbf288)


Note that the login process reflects organizations, not projects,
because this entire UI isn't meant to work for the dogfood use-case:


![screenshot-2025-01-24-15-47-53](https://github.com/user-attachments/assets/76aea057-5dd3-47fd-8811-fd9ddeb2aac1)

(Incoherent things happen if you create an organization using the
intermediate API in our project.)