User impersonation (#171)
Closes OPEN-145, OPEN-146, OPEN-147, OPEN-148, OPEN-149.

User impersonation works by letting owners of a project issue tokens.
Those tokens are tied to a user in their project, and also to
themselves. A new endpoint in intermediate can exchange that token for a
session on the target user.

This PR also includes a button in the app to create user impersonation
tokens, which redirects to a new page in the portal that automatically
redeems them.


![screenshot-2025-01-26-10-36-27](https://github.com/user-attachments/assets/c7438e06-0b18-4962-83fd-09a7e4b2c8ee)

Tesseral customers can tell if an access token is from impersonation
from a new claim on the access token:

```json
{
  "iss": "https://project-54vwf0clhh0caqe20eujxgpeq.tesseral.app",
  "sub": "user_dt7bvnh0qnn5qhej0nclbb04b",
  ...
  "impersonator": {"email":"root@app.tesseral.example.com"},
}
```

The intention of this claim is:

1. If customers have certain endpoints they don't want called by
impersonated users, they can detect and disable such invocations. Just
check if the `impersonator` claim is present.
2. On customer's apps, they can display a banner reminding them that
they're not logged in as themselves, but instead as a customer. The
intent of including details about the impersonator on the claims is so
that these banners can actively say things like "Remember John Doe, your
actions are being audited."

I'm open to just moving to an `impersonated: true` claim. Genuinely
unsure.