[OPEN-114] Make organization auth overrides more fail-safe (#173)
The current columns for overriding auth methods at the org level are
prone to error. They say "override", but they don't have unilateral
ability to enable an auth method. If the project has an auth method
disabled, that always disables it for the org.

This PR has the columns be called disable_log_in_with_password. The
meaning of NULL is unchanged. The meaning for non-null values is
flipped. It's much harder to accidentally think that column has meaning
on its own.

This also simplifies a lot of logic, where now the right behavior is a
bunch of code like:

```
enabled := project.enabled
if derefOrEmpty(org.disabled) {
  enabled = false
}
```

See the diff to see how much simpler this is than the previous design.

I opt to name the CHECK constraint
("no_disable_logins_unless_override_log_in_methods") so that it can be
ALTERed when we add more supported login methods.