[OPEN-165] Add UI for organization Google hosted domains, Microsoft tenant IDs (#181)
This PR adds a UI for viewing and editing the Google HDs and Microsoft
TIDs on an organization. To make this work, I had to unbreak some
unrelated stuff in main.

InputTags is mostly copied over from SSOReady. It lets you edit a list
of tag-like properties (i.e. unordered, no spaces, feels conceptually OK
to display in a badge). I use that to let you control the set of field
values.


![screenshot-2025-01-29-13-57-39](https://github.com/user-attachments/assets/d271c547-1760-416b-879e-dd712df04d43)

![screenshot-2025-01-29-13-57-35](https://github.com/user-attachments/assets/e4ded90b-4ad9-4b47-9a44-8292f0dbd81b)

![screenshot-2025-01-29-13-57-31](https://github.com/user-attachments/assets/8519cdb0-3bdf-4270-80db-48f3cfcc5299)