[OPEN-168] complete authenticator verification (#202)
This PR finishes the verification flow (logging in) for authenticator
apps.

It also resolves some issues with the routing logic for secondary
factors to ensure that we only route to the `ChooseAdditionalFactor`
view if the user either has no secondary factors registered or has
multiple registered. In the cases where they only have a single
secondary factor registered, we display the single factor's verification
view instead.