[OPEN-202] Add `log_in_with_email` and `log_in_with_saml` to database and backend API (#226)
This PR adds `log_in_with_email` and `log_in_with_saml` to the database
models. It also has the backend API support CRUDing these properties,
which behave like all other login methods for that purpose.

As part of this PR, I made minimal changes to the frontend and
intermediate APIs. That's just to make things compile. Follow-on PRs
will change these.

This PR does not update the console, which is still expecting
`saml_enabled` and which doesn't know about "log in with email" at all.