[OPEN-218] Add list of trusted domains to projects, replace passkey RP IDs (#251)
This PR drops the concept of "passkey RP IDs" on projects. Instead,
projects now have a list of trusted domains. Most of this PR essentially
just renames the concept.

One difference from before is that the project's default vault domain
(the .tesseral.app one) is unconditionally included in the set of
trusted domains. This was not the case for the passkey RP IDs.