[OPEN-217] add console sidebar (#249)
This PR adds a basic sidebar to the console pages. For now it's just a
dark sidebar, but this will be restyled and the items will be finalized
at some point in the future. This is mostly just to get it in there.

<img width="1660" alt="Screenshot 2025-02-07 at 6 20 04 PM"
src="https://github.com/user-attachments/assets/329660ce-af5f-4823-915f-316c5b3586c7"
/>