Add UpdatePasskey to backend, console (#259)
This PR adds an UpdatePasskey endpoint to the backend API. It's only
useful for disabling and enabling passkeys. The console adds this
functionality to the danger zone.

I also opt to change how we display the credential ID of a passkey. It's
now hex-encoded, since that's much more likely to be useful for folks
who want to quickly test out a passkey invocation when making their own
UI. Hex is less ambiguous in practice, and easier to recognize by a
developer.


![screenshot-2025-02-11-11-16-23](https://github.com/user-attachments/assets/61042207-8723-4615-919b-f2e82989a8d1)