Add redirect_uri, after_login/signup_redirect_uri project properties (#258)
This PR dumbs down our project redirect system. It's now just a trio of
plain strings on a project.

```json
{
  "project": {
    "id": "project_54vwf0clhh0caqe20eujxgpeq",
    "displayName": "Tesseral Local Development",
    "createTime": "2025-02-07T19:54:09.702411Z",
    "updateTime": "2025-02-11T18:19:00.034667Z",
    "logInWithGoogle": true,
    "logInWithMicrosoft": true,
    "logInWithEmail": true,
    "logInWithPassword": true,
    "logInWithSaml": true,
    "logInWithAuthenticatorApp": true,
    "logInWithPasskey": true,
    "authDomain": "auth.console.tesseral.example.com",
    "trustedDomains": [
      "auth.console.tesseral.example.com",
      "project-54vwf0clhh0caqe20eujxgpeq.tesseral.example.app",
      "console.tesseral.example.com"
    ],
    "redirectUri": "https://console.tesseral.example.com",
    "afterLoginRedirectUri": "https://console.tesseral.example.com/after-login",
    "afterSignupRedirectUri": "https://console.tesseral.example.com/after-signup"
  }
}
```

We anticipate that later, folks will want to be able to redirect to a
vault and then back to a specific page. We can figure that out when we
have a concrete use-case at hand.

For now, the intention is to have the redirect_uri be the
general-purpose place to send end users back to from the vault. Two
separate properties special-case the redirect behavior after signing in
or signing up, but the vault will default to the redirect_uri if that
property is not set.

If none of these properties are set? I reckon we redirect to the vault
user settings page, and have the onboarding flow and console discourage
such a setup.