Fix console session management (#272)
This PR updates the `useAccessToken` hook to use an old fashioned
`fetch` and persist pending status in a state variable. This seems to
resolve the crazy reload bug we were seeing.

Also, while I was in there I noticed that our refresh token cookies were
only session cookies, so I updated it to have them be 1 year cookies.