Update prettier and eslint for console pages (#271)
This PR updates the prettier and eslint configs to use updated standards
and applies fixes to resolve errors.

Note: currently there's an issue with `react-hook-form` that appears to
be resolved in an upcoming release, so I've disabled the eslint rule for
form declaration and usage statements.

These will be removed once `7.55.0` of `react-hook-form` is released.