Verify emails using links (#299)
There's ways to mitigate stuffing attacks, but the simpler option is to
forego this stuff. Instead, make email verification codes have ~128 bits
of entropy.

Downside is we have to re-do the UI to be based on links (with a
fallback to copy-paste). This PR does that. We also lose the nicety of
"open email on my phone, type it on my desktop". I accept this.


![screenshot-2025-02-28-11-08-55](https://github.com/user-attachments/assets/cc01c354-7293-4304-8898-19d3bbe76471)

![screenshot-2025-02-28-11-08-57](https://github.com/user-attachments/assets/080ccc25-35c9-4513-8d04-9ace8b393f1b)

![screenshot-2025-02-28-11-08-59](https://github.com/user-attachments/assets/d97974a2-db90-4ab4-8ded-1f059b5bc341)