Add basics for organization switching (#319)
This PR adds the basics for organization switching: a frontend endpoint
for listing switchable organizations, and an intermediate endpoint that
exchanges sessions for intermediate sessions.

This is only the "basics" because the login flow cannot easily
accommodate the steps that follow choosing an organization from this
logic. Doing so would require a lot of copy-pasting. Instead, a later PR
will rewrite our login flow to be based on URLs and links, instead of
the current "view" concept and `deriveNextView` logic.