Rework vaultui to support switching orgs (#320)
I couldn't think of a better way to do this than to just move the login
flow to use URL-based schemes, with a centralized hook that does the
appropriate next redirect. It was hard to reason about what's loaded or
not with the current setup.

This PR does not recover organization settings, but we can add it back.

This PR leaves the empty path blank. I was thinking we dynamically
choose whether to redirect to user settings or to the login flow based
on whether the user is logged in or not?

A follow-on PR needs to re-add the sidebar to the vault. Said sidebar
needs to implement switching organizations; that way, Tesseral customers
who don't want to implement organization switching UI at all can just
piggyback off the vault.