[OPEN-268] Make authenticator app recovery codes uuidv4s (#333)
(Draft PR, will rebase on top of upcoming work to unify console and
vault login flow code)


![screenshot-2025-03-25-14-29-07](https://github.com/user-attachments/assets/dfc216c6-c547-4863-bd04-d10585f51a94)


![screenshot-2025-03-25-14-38-04](https://github.com/user-attachments/assets/5b60ab61-578b-496a-9211-eb9a8979c9d8)

This PR makes authenticator apps use UUIDv4s for authenticator app
recovery codes. This makes them quite annoying to handle when the
clipboard isn't available, but much more secure. So much so that we can
use SHA256 instead of bcrypt to store and validate them.

In the process of making this happen, I:

1. Added authenticator app registration to the vault dashboard
2. Made style changes to how we display the authenticator codes
3. Fixed some backend logic that failed to find the organization the
user belongs to when registering authenticator apps. The current logic
only worked for the dogfood project.