Add login/signup flow to console (#337)
This is almost entirely a copy-paste from the vault. The import paths
even worked as-is. The main changes were to the SignupPage (reskin),
LoginPage (same), and CreateOrganizationPage (hooked up to new
OnboardingCreateProjects endpoint). A few places I renamed
"organizations" to "projects".

OnboardingCreateProjects is a bit exciting. It's a sui generis endpoint
that creates two projects (dev, prod) and logs the caller into the dev
project. Its associated UI is a bit brutalist, but might be good enough
for initial launch to friendlies?

A follow-on PR will gut a lot of the console's logic for keeping track
of its own login state. It can be greatly simplified, because we can
always assume a cookie is present.

Also, I removed the leading dot from the cookies we attempt to write. Go
silently strips them out, for good reason:
https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/Set-Cookie#domaindomain-value


![screenshot-2025-03-26-15-50-46](https://github.com/user-attachments/assets/5f6d9adc-3781-4190-8041-8ad15e7f7f11)

![screenshot-2025-03-26-15-50-49](https://github.com/user-attachments/assets/e70a4b2b-6b47-4658-8cae-ffec79bb6559)

![screenshot-2025-03-26-15-51-02](https://github.com/user-attachments/assets/d7f4b7e0-36b6-473d-8118-7a6974865884)

![screenshot-2025-03-26-15-51-05](https://github.com/user-attachments/assets/b3f588a7-68cf-4018-9084-688276c10b23)