Fix relayed session redirect, update project domains, frontend Bearer auth (#339)
Sorry for the formatter noise in this PR.

This PR fixes the redirect we do for relayed sessions. 

It also fixes a frontend bug where we issue an update to cookie domain,
which is immutable until customized. I opt to avoid touching it from the
frontend.

I'm also including a change that allows calls to the frontend API from a
Bearer token. This is more convenient for mobile users.