[OPEN-346] Create Users and Organizations from the console (#376)
This PR adds support for creating Organizations and Users from the
console. Given the restrictions of the types of Organizations that can
be associated with the Dogfood Project, there is also a check added to
the `CreateOrganization` function in the backend store to prevent this
RPC from being run on the Dogfood Project.

![Screenshot 2025-04-14 at 12 47
52 PM](https://github.com/user-attachments/assets/bf04ad1f-43a3-4f87-848c-6ca5a10b0ffc)
![Screenshot 2025-04-14 at 12 47
59 PM](https://github.com/user-attachments/assets/5cc02af7-fbeb-4743-a3a4-add21ebaa4f2)
![Screenshot 2025-04-14 at 12 48
10 PM](https://github.com/user-attachments/assets/5443dbb4-cf1c-45ad-aefe-b2a45778e5fc)
![Screenshot 2025-04-14 at 12 48
15 PM](https://github.com/user-attachments/assets/9809e358-7039-4b46-b174-6ba4ad365238)