[OPEN-362] Resolve React cookie management issues (#382)
This PR resolves issues with cookie management in React (the same issue
we saw with [the
SDK](https://github.com/tesseral-labs/tesseral-sdk-react/pull/4)). We're
now holding returned access tokens in memory to get around the
unreliability of document.cookie in the React lifecycle.

There was also an issue with some of the `App.tsx` imports and routes
since Google and Microsoft settings are in modals now, so I resolved
those to get the local build working.