[OPEN-384] Add `auto_create_organizations` option to Vault UI Settings (#389)
This PR adds support for a signup flow that doesn't mention
Organizations at all. Instead, an Organization is auto-created based on
the User's display name or email.

To do this, this PR:

1. Adds `project_ui_settings.auto_create_organizations` and its
associated CRUD.
2. Adds to the existing logic on ListOrganizationsPage that "shortcuts"
to CreateOrganizationPage, and instead auto-creates the organization.

This PR also makes some modifications to nginx config, to handle some
race conditions in local development that I ran into. I haven't gotten
to the bottom of it, but I can confirm that I was in a state where all
of the modifications I made to nginx configuration were required in
concert.


![screenshot-2025-05-07-10-41-49](https://github.com/user-attachments/assets/fdf86f62-243d-46c8-9b01-eb1370f5379f)


![screenshot-2025-05-07-10-41-31](https://github.com/user-attachments/assets/9322b819-c59b-4cde-9e72-c17fbe87f3ce)


![screenshot-2025-05-07-10-41-21](https://github.com/user-attachments/assets/293aeb90-b0aa-4c23-8a97-cf9b120d2041)