[OPEN-379] [OPEN-380] [OPEN-381] [OPEN-382] Webhook creation and management (#391)
This PR updates the `createProjectForCurrentUser` procedure to include a
call out to Svix to create an Application to handle Webhook traffic for
the created Project. It also adds a new card to the Project details view
in the console to allow management of the Project's Webhook. The `Edit`
button simply kicks out to Svix for Webhook settings management via the
URL provided by their `svixClient.Authentication.AppPortalAccess()`
function.

![Screenshot 2025-05-08 at 12 33
31 PM](https://github.com/user-attachments/assets/a4b2ab30-154e-4cca-b1d4-742df97b3467)

![Screenshot 2025-05-08 at 12 34
03 PM](https://github.com/user-attachments/assets/197299ca-2283-498d-86bb-914e2cd1b745)

It also adds event dispatching during the creation, updating, and
deletion of users and organizations for all services.

![Screenshot 2025-05-08 at 2 46
42 PM](https://github.com/user-attachments/assets/4f94a572-b2cf-4d71-8820-9a7c87f04b8c)