[OPEN-369] Role-Based Access Control (#396)
This PR adds managed Role-Based Access Control to Projects. Tesseral
customers can now:

1. Define a set of fine-grained actions that users can take in their
product.
2. Define a set of roles available to all organizations their project.
Users can be assigned to roles, and roles are basically just a named
collection of actions.
3. Tesseral customers can optionally enable custom roles for an
organization, allowing owners of that organization to define custom
roles on top of the actions from (1).
4. When a user has a role, the corresponding actions are listed in a new
`actions` claim on access tokens. The SDKs will have convenience methods
for looking up the contents of this claim.

Managing users, roles, and user-to-role assignments are all available
self-serve in the Vault UI. Tesseral customers do not need to make any
code changes.

An example of the new `actions` claim:

```json
{
  "actions": [
    "acme.workspaces.view"
  ],
  "aud": "https://project-e1md0tokbplxhynmvs219re00.tesseral.app",
  "exp": 1747155294,
  "iat": 1747154994,
  "impersonator": {
    "email": "ulysse.carion@ssoready.com"
  },
  "iss": "https://project-e1md0tokbplxhynmvs219re00.tesseral.app",
  "nbf": 1747154994,
  "organization": {
    "displayName": "FooBar",
    "id": "org_3xslzq7n9043yzrle4sh60moo"
  },
  "session": {
    "id": "session_03dx4q1vzgpbxene3i2t5nwxi"
  },
  "sub": "user_24pd1g7k2wqo6jl9eg7y9va99",
  "user": {
    "email": "asdf@foo.com",
    "id": "user_24pd1g7k2wqo6jl9eg7y9va99"
  }
}
```


![screenshot-2025-05-13-09-50-58](https://github.com/user-attachments/assets/42ae76e1-587c-40eb-a6c6-c0ced2fc41c0)


![screenshot-2025-05-13-09-51-10](https://github.com/user-attachments/assets/17b6f636-0cc1-4ad5-b2a8-d88c0c8e0d53)