Simplify nginx configuration and force lazy DNS resolution (#406)
This PR makes changes only for local development. In particular, this PR
changes how nginx configuration works in local dev:

1. All nginx server blocks that front other docker containers are
configured with `resolver 127.0.0.11 valid=1s;` and do not directly
`proxy_pass http://container;`, instead doing a `set $upstream
http://container; proxy_pass $upstream;`. This has the effect of forcing
nginx to do DNS lookups at HTTP request time, and of re-resolving
container addresses with a 1-second TTL. This TTL-ing is necessary
because if the proxied container is restarted, it may be come back on a
different IP address.
2. Most of the parameters we were passing to `proxy_pass` are
unnecessary. Simple is easier to fix down the line. IPv6 configuration
also got the same treatment.
3. I reorganized the directives to always start with the `server_name`,
because they tell you what you're looking at.

I can't find first-party documentation on the variable-related behavior
in (1), but there many references to this behavior in the wild.