Add redirect_uri and return_relayed_session_token_as_query_param (#409)
This PR adds two new options on intermediate sessions, which can be
populated from links to the vault `/login` page:

* `redirect_uri` must be from the project's set of trusted domains, and
is a URL to redirect to instead of the project-configured URI.
* `return_relayed_session_token_as_query_param` indicates that relayed
session tokens should be returned in a query parameter instead of a
fragment.

This functionality is added in service of web frameworks which expect
session management to be performed server-side on specialized endpoints,
starting with Next.js.