Fix logout with bad token
When logging out with a bad access token, the Logout endpoint fails with a 401 unauthorized code.
However, the access token and refresh token remain as cookies since they are not cleared by the Logout endpoint.

This can be easily seen in dev by logging in, re-bootstrapping, refreshing the console, and trying to logout.
Nothing happens because the logout call is silently failing.