Fix email input in login page (#447)
The email input field was missing `type="email"` which led mobile
browsers to automatically capitalize the first letter of the email
address. This in turn led to errors logging in because the users' emails
are case-sensitive. This change ensures that the input field is treated
as an email input, preventing the issue.

Confirmed fix via iOS Simulator.