Fix spurious logouts in console
The main issue was that `document.cookie` does not return expired cookies,
so in the case an access token was expired but the refresh token was still valid,
the console would redirect to the login page while a refresh was in progress.

This ensures that nothing is rendered until an initial refresh is completed,
in the case that the access token is expired or not present.

Since the refresh token is HTTP-only, attempting a refresh is the only way
to know if one is available at all.