[OPEN-451] Match on GithubUserID during exchange (#461)
During the process of exchanging an Intermediate Session for a full
Session, we perform a matching step to determine if a user already
exists with the same credential (Google User ID, Microsoft User ID,
email, etc.).

Currently, we're not including GitHub User ID in this matching step,
which means GitHub logins will always fall back to email when
determining if a user should be created or not.

This is less than ideal for a number of reasons:
- It strays from the established exchange flow for other OAuth providers
- If a user changes their primary GitHub email address, they'll no
longer be able to create a session for their existing user

---

This PR adds GitHub User ID matching to the exchange step to resolve the
above issues.