Allow editing of RBAC actions (#463)
Currently, once an action is added to the RBAC policy, its description
is effectively permanent.

The way that the current `UpsertAction` query works, it simply ignores
description changes when performing the Upsert.

---

This PR updates the `UpsertAction` query in the backend service to
update the description on conflicts, allowing for editing actions from
the Console and API.