[OPEN-417] Simplify email + password flow (#455)
Today, "Log in with Password" is essentially an email magic link flow,
followed by a password verification step. This PR changes that flow to
support providing both email and password on login, simplifying the
process.

<img width="435" alt="Screenshot 2025-06-09 at 2 36 15 PM"
src="https://github.com/user-attachments/assets/0ede27a4-b47a-4d83-9eee-805a47e28155"
/>

This PR also fixes a bug wherein password registration is only correctly
supported on new users. This causes problems if:

1. An end user signs up using "Log in with Email".
2. Later, that end user signs up with "Log in with Password". This can
most readily happen if the Organization disables "Log in with Email" in
the meantime.

Today, our behavior is to require password registration, but our logic
for exchanging an Intermediate Session for a (real) Session fails to
update the User with the newly registered password. This PR fixes that,
and now registered passwords are properly copied to the User.

# Context

Tesseral's login system upholds the following invariants:

1. Users have verified emails. There is no such thing as a
`email_verified` property on a User.
4. The existence of Users and Organizations is confidential by default.
To know that a User exists in an Organization, you must first verify
control of that User's email.
5. It's valid for there two be many Users with the same email, but in
different Organizations. These Users are governed by the Organization
they're owned by, and have distinct passwords.

The "confidential by default" rule is to deter attempts at probing for
the presence of a customer. This objective is not always attainable, if
an Organization uses auth mechanisms that don't support such
confidentiality. For example, enabling Log in with SAML means that
Organization's presence can be detected by inputting an email from that
Organization. This is ubiquitous in B2B SaaS, because it's a
prerequisite for SP-initiated SAML flows.

For the sake of parsimony, our first implementation of Log in with
Password relegated passwords to a secondary factor, because:

1. Password-based signups need to verify their email address in the
signup flow, because there can never be a User with an "unverified"
email.
2. Password-based logins take an (unverified) email and password, and
need to know which User to try the password against; there may be
multiple, each in a different Organization. We certainly aren't going to
allow you to try the same password against all such Users in parallel.
We also can't offer you the list of Organizations you might be referring
to, because we keep that information confidential until you verify your
email.

These issues were simplest to resolve by having passwords happen after
email verification.

# The new flow

The above constraints are not possible to avoid in full generality, but
a happy path exists that we can service more easily:

* During the login flow, we can support an email + password input if the
given email uniquely identifies a single, password-having User. In that
flow, we can check the email/password pair against a single user in the
usual way (i.e. with password stuffing protection, etc.), and if that
check passes, then we will consider the login attempt to have
successfully verified both its email and password.
* In the case of there not being multiple password-having Users with
that email, then we can fall back to first requiring email verification.
From there, the end user can choose which User to log into, then input a
password.
* The same fallback flow also works in the case of there being zero
existing Users with that email. This effectively becomes the same thing
as the password-based signup flow.

In the case of end users having many Organizations, such end users can
consistently hit the happy path of unique email by registering each user
with a distinct email address. For example, if the end user is on Gmail
or Google Workspace, they can create users
`john.doe+acme-dev@example.com`, `john.doe+acme.stage@example.com`, etc.
This pattern too is ubiquitous in B2B SaaS, for example with the emails
for root AWS accounts.

This change represents an amendment to our User presence confidentiality
posture. When a Project enables "Log in with Password" and its
Organizations keep it enabled, then User presence is not confidential.
We are very comfortable with such an amendment, because this is
ubiquitous in B2B SaaS. Systems supporting email + password logins also
support email-presence probing, either from error states (a different
error for bad email vs bad password), or ultimately from timing attacks
(you can't do a bcrypt check if there is no bcrypt, and bcrypt has very
noticeable timing effects).